+++ app/src/src_nonfreenet/org/breezyweather/sources/openweather/OpenWeatherResultConverter.kt
@@ -18,17 +18,10 @@ package org.breezyweather.sources.openweather
 
 import breezyweather.domain.location.model.Location
 import breezyweather.domain.weather.model.AirQuality
-import breezyweather.domain.weather.model.Alert
-import breezyweather.domain.weather.model.AlertSeverity
-import breezyweather.domain.weather.model.Astro
-import breezyweather.domain.weather.model.Current
 import breezyweather.domain.weather.model.Daily
-import breezyweather.domain.weather.model.HalfDay
-import breezyweather.domain.weather.model.Minutely
 import breezyweather.domain.weather.model.Precipitation
 import breezyweather.domain.weather.model.PrecipitationProbability
 import breezyweather.domain.weather.model.Temperature
-import breezyweather.domain.weather.model.UV
 import breezyweather.domain.weather.model.WeatherCode
 import breezyweather.domain.weather.model.Wind
 import breezyweather.domain.weather.wrappers.AirQualityWrapper
@@ -37,113 +30,79 @@ import breezyweather.domain.weather.wrappers.SecondaryWeatherWrapper
 import breezyweather.domain.weather.wrappers.WeatherWrapper
 import org.breezyweather.common.exceptions.InvalidOrIncompleteDataException
 import org.breezyweather.common.extensions.capitalize
+import org.breezyweather.common.extensions.getFormattedDate
 import org.breezyweather.common.extensions.toDate
-import org.breezyweather.common.extensions.toTimezoneNoHour
+import org.breezyweather.common.extensions.toDateNoHour
 import org.breezyweather.sources.getDailyAirQualityFromHourly
 import org.breezyweather.sources.openweather.json.OpenWeatherAirPollution
 import org.breezyweather.sources.openweather.json.OpenWeatherAirPollutionResult
-import org.breezyweather.sources.openweather.json.OpenWeatherOneCallAlert
-import org.breezyweather.sources.openweather.json.OpenWeatherOneCallDaily
-import org.breezyweather.sources.openweather.json.OpenWeatherOneCallHourly
-import org.breezyweather.sources.openweather.json.OpenWeatherOneCallMinutely
-import org.breezyweather.sources.openweather.json.OpenWeatherOneCallResult
+import org.breezyweather.sources.openweather.json.OpenWeatherForecast
+import org.breezyweather.sources.openweather.json.OpenWeatherForecastResult
 import java.util.Date
-import java.util.Objects
-import kotlin.math.roundToInt
 
 fun convert(
     location: Location,
-    oneCallResult: OpenWeatherOneCallResult,
+    forecastResult: OpenWeatherForecastResult,
     airPollutionResult: OpenWeatherAirPollutionResult?
 ): WeatherWrapper {
-    // If the API doesn’t return hourly or daily, consider data as garbage and keep cached data
-    if (oneCallResult.hourly.isNullOrEmpty() || oneCallResult.daily.isNullOrEmpty()) {
+    // If the API doesn’t return hourly, consider data as garbage and keep cached data
+    if (forecastResult.list.isNullOrEmpty()) {
         throw InvalidOrIncompleteDataException()
     }
 
     val hourlyAirQuality = getHourlyAirQuality(airPollutionResult?.list)
 
     return WeatherWrapper(
-        /*base = Base(
-            publishDate = oneCallResult.current?.dt?.times(1000)?.toDate() ?: Date()
-        ),*/
-        current = if (oneCallResult.current != null) Current(
-            weatherText = oneCallResult.current.weather?.getOrNull(0)?.description?.capitalize(),
-            weatherCode = getWeatherCode(oneCallResult.current.weather?.getOrNull(0)?.id),
+        /*current = if (currentResult != null) Current(
+            weatherText = forecastResult.current.weather?.getOrNull(0)?.description?.capitalize(),
+            weatherCode = getWeatherCode(forecastResult.current.weather?.getOrNull(0)?.id),
             temperature = Temperature(
-                temperature = oneCallResult.current.temp,
-                apparentTemperature = oneCallResult.current.feelsLike
+                temperature = forecastResult.current.temp,
+                apparentTemperature = forecastResult.current.feelsLike
             ),
             wind = Wind(
-                degree = oneCallResult.current.windDeg?.toDouble(),
-                speed = oneCallResult.current.windSpeed,
-                gusts = oneCallResult.current.windGust
+                degree = forecastResult.current.windDeg?.toDouble(),
+                speed = forecastResult.current.windSpeed,
+                gusts = forecastResult.current.windGust
             ),
-            uV = UV(index = oneCallResult.current.uvi),
-            relativeHumidity = oneCallResult.current.humidity?.toDouble(),
-            dewPoint = oneCallResult.current.dewPoint,
-            pressure = oneCallResult.current.pressure?.toDouble(),
-            cloudCover = oneCallResult.current.clouds,
-            visibility = oneCallResult.current.visibility?.toDouble()
-        ) else null,
-        dailyForecast = getDailyList(oneCallResult.daily, hourlyAirQuality, location),
-        hourlyForecast = getHourlyList(oneCallResult.hourly, hourlyAirQuality),
-        minutelyForecast = getMinutelyList(oneCallResult.minutely),
-        alertList = getAlertList(oneCallResult.alerts)
+            uV = UV(index = forecastResult.current.uvi),
+            relativeHumidity = forecastResult.current.humidity?.toDouble(),
+            dewPoint = forecastResult.current.dewPoint,
+            pressure = forecastResult.current.pressure?.toDouble(),
+            cloudCover = forecastResult.current.clouds,
+            visibility = forecastResult.current.visibility?.toDouble()
+        ) else null,*/
+        dailyForecast = getDailyList(forecastResult.list, hourlyAirQuality, location),
+        hourlyForecast = getHourlyList(forecastResult.list, hourlyAirQuality)
     )
 }
 
 private fun getDailyList(
-    dailyResult: List<OpenWeatherOneCallDaily>,
+    hourlyResult: List<OpenWeatherForecast>,
     hourlyAirQuality: MutableMap<Date, AirQuality>,
     location: Location
 ): List<Daily> {
     val dailyAirQuality = getDailyAirQualityFromHourly(hourlyAirQuality, location)
-    val dailyList: MutableList<Daily> = ArrayList(dailyResult.size)
-    for (i in 0 until dailyResult.size - 1) {
-        val dailyForecast = dailyResult[i]
-        val theDay = Date(dailyForecast.dt.times(1000)).toTimezoneNoHour(location.javaTimeZone)!!
-        dailyList.add(
-            Daily(
-                date = theDay,
-                day = HalfDay(
-                    weatherText = dailyForecast.weather?.getOrNull(0)?.description?.capitalize(),
-                    weatherPhase = dailyForecast.weather?.getOrNull(0)?.description,
-                    weatherCode = getWeatherCode(dailyForecast.weather?.getOrNull(0)?.id),
-                    temperature = Temperature(
-                        temperature = dailyForecast.temp?.max,
-                        apparentTemperature = dailyForecast.feelsLike?.eve
-                    )
-                ),
-                night = HalfDay(
-                    weatherText = dailyForecast.weather?.getOrNull(0)?.description?.capitalize(),
-                    weatherPhase = dailyForecast.weather?.getOrNull(0)?.description,
-                    weatherCode = getWeatherCode(dailyForecast.weather?.getOrNull(0)?.id),
-                    // night temperature is actually from previous night,
-                    // so we try to get night from next day if available
-                    temperature = Temperature(
-                        temperature = dailyResult[i + 1].temp?.min,
-                        apparentTemperature = dailyResult[i + 1].feelsLike?.morn
-                    )
-                ),
-                sun = Astro(
-                    riseDate = dailyForecast.sunrise?.times(1000)?.toDate(),
-                    setDate = dailyForecast.sunset?.times(1000)?.toDate()
-                ),
-                moon = Astro(
-                    riseDate = dailyForecast.moonrise?.times(1000)?.toDate(),
-                    setDate = dailyForecast.moonset?.times(1000)?.toDate()
-                ),
-                airQuality = dailyAirQuality.getOrElse(theDay) { null },
-                uV = UV(index = dailyForecast.uvi)
+    val dailyList = mutableListOf<Daily>()
+    val hourlyListByDay = hourlyResult.groupBy {
+        it.dt.times(1000).toDate().getFormattedDate("yyyy-MM-dd", location)
+    }
+    for (i in 0 until hourlyListByDay.entries.size - 1) {
+        val dayDate = hourlyListByDay.keys.toTypedArray()[i].toDateNoHour(location.javaTimeZone)
+        if (dayDate != null) {
+            dailyList.add(
+                Daily(
+                    date = dayDate,
+                    airQuality = dailyAirQuality.getOrElse(dayDate) { null }
+                )
             )
-        )
+        }
     }
     return dailyList
 }
 
 private fun getHourlyList(
-    hourlyResult: List<OpenWeatherOneCallHourly>,
+    hourlyResult: List<OpenWeatherForecast>,
     hourlyAirQuality: MutableMap<Date, AirQuality>
 ): List<HourlyWrapper> {
     return hourlyResult.map { result ->
@@ -153,26 +112,24 @@ private fun getHourlyList(
             weatherText = result.weather?.getOrNull(0)?.main?.capitalize(),
             weatherCode = getWeatherCode(result.weather?.getOrNull(0)?.id),
             temperature = Temperature(
-                temperature = result.temp,
-                apparentTemperature = result.feelsLike
+                temperature = result.main?.temp,
+                apparentTemperature = result.main?.feelsLike
             ),
             precipitation = Precipitation(
-                total = getTotalPrecipitation(result.rain?.cumul1h, result.snow?.cumul1h),
-                rain = result.rain?.cumul1h,
-                snow = result.snow?.cumul1h
+                total = getTotalPrecipitation(result.rain?.cumul3h, result.snow?.cumul3h),
+                rain = result.rain?.cumul3h,
+                snow = result.snow?.cumul3h
             ),
             precipitationProbability = PrecipitationProbability(total = result.pop?.times(100.0)),
             wind = Wind(
-                degree = result.windDeg?.toDouble(),
-                speed = result.windSpeed,
-                gusts = result.windGust
+                degree = result.wind?.deg?.toDouble(),
+                speed = result.wind?.speed,
+                gusts = result.wind?.gust
             ),
             airQuality = hourlyAirQuality.getOrElse(theDate) { null },
-            uV = UV(index = result.uvi),
-            relativeHumidity = result.humidity?.toDouble(),
-            dewPoint = result.dewPoint,
-            pressure = result.pressure?.toDouble(),
-            cloudCover = result.clouds,
+            relativeHumidity = result.main?.humidity?.toDouble(),
+            pressure = result.main?.pressure?.toDouble(),
+            cloudCover = result.clouds?.all,
             visibility = result.visibility?.toDouble()
         )
     }
@@ -186,42 +143,6 @@ private fun getTotalPrecipitation(rain: Double?, snow: Double?): Double? {
     return if (snow == null) {
         rain
     } else rain + snow
-}
-
-private fun getMinutelyList(minutelyResult: List<OpenWeatherOneCallMinutely>?): List<Minutely> {
-    val minutelyList: MutableList<Minutely> = arrayListOf()
-    minutelyResult?.forEachIndexed { i, minutelyForecast ->
-        minutelyList.add(
-            Minutely(
-                date = Date(minutelyForecast.dt * 1000),
-                minuteInterval = if (i < minutelyResult.size - 1) {
-                    ((minutelyResult[i + 1].dt - minutelyForecast.dt) / 60).toDouble().roundToInt()
-                } else ((minutelyForecast.dt - minutelyResult[i - 1].dt) / 60).toDouble()
-                    .roundToInt(),
-                precipitationIntensity = minutelyForecast.precipitation
-            )
-        )
-    }
-    return minutelyList
-}
-
-private fun getAlertList(resultList: List<OpenWeatherOneCallAlert>?): List<Alert> {
-    return if (resultList != null) {
-        return resultList.map { result ->
-            Alert(
-                // Create unique ID from: alert event, start time
-                alertId = Objects.hash(result.event, result.start).toString(),
-                startDate = Date(result.start.times(1000)),
-                endDate = Date(result.end.times(1000)),
-                headline = result.event,
-                description = result.description,
-                source = result.senderName,
-                severity = AlertSeverity.UNKNOWN // Does not exist
-            )
-        }
-    } else {
-        emptyList()
-    }
 }
 
 private fun getWeatherCode(icon: Int?): WeatherCode? {
@@ -265,18 +186,11 @@ private fun getHourlyAirQuality(
 }
 
 fun convertSecondary(
-    oneCallResult: OpenWeatherOneCallResult?,
     airPollutionResult: OpenWeatherAirPollutionResult?
 ): SecondaryWeatherWrapper {
     return SecondaryWeatherWrapper(
         airQuality = if (airPollutionResult != null) {
             AirQualityWrapper(hourlyForecast = getHourlyAirQuality(airPollutionResult.list))
-        } else null,
-        minutelyForecast = if (oneCallResult != null) {
-            getMinutelyList(oneCallResult.minutely)
-        } else null,
-        alertList = if (oneCallResult != null) {
-            getAlertList(oneCallResult.alerts)
         } else null
     )
 }