+++ app/src/src_nonfreenet/org/breezyweather/sources/openweather/json/OpenWeatherForecastPrecipitation.kt
@@ -20,21 +20,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class OpenWeatherOneCallHourly(
-    val dt: Long,
-    val temp: Double?,
-    @SerialName("feels_like") val feelsLike: Double?,
-    val pressure: Int?,
-    val humidity: Int?,
-    @SerialName("dew_point") val dewPoint: Double?,
-    val uvi: Double?,
-    val clouds: Int?,
-    val visibility: Int?,
-    @SerialName("wind_speed") val windSpeed: Double?,
-    @SerialName("wind_deg") val windDeg: Int?,
-    @SerialName("wind_gust") val windGust: Double?,
-    val weather: List<OpenWeatherOneCallWeather>?,
-    val pop: Double?,
-    val rain: OpenWeatherOneCallPrecipitation?,
-    val snow: OpenWeatherOneCallPrecipitation?
+data class OpenWeatherForecastPrecipitation(
+    @SerialName("3h") val cumul3h: Double?
 )